Declare magit-key-mode-groups with (defvar ...)
Byte compiler doesn't like assignments to constants.
And anyway this isn't a constant.

The one difference (defconst...) makes is it restores the original
value upon reloading the file, but if that's what you want, then you'd
also need to reset magit-key-mode-key-maps as well, meaning that, too,
should also be a defconst.  Which would then give us two nonconstant
constants.  Better to do

    (defvar magit-key-mode-key-maps '() "This will be filled lazily...")
    ...
    (defvar magit-key-mode-groups '()  "Holds the key, help, ...")

    (defconst magit-key-mode-default-groups
     '((logging ...) <rest of original value for magit-key-mode-groups>)
     "Default value for magit-key-mode-groups (which see).")

    (defun magit-restore-key-mode-groups ()
      (setq magit-key-mode-groups magit-key-mode-default-groups)
      (setq magit-key-mode-key-maps nil))

    (magit-restore-key-mode-groups)