move magit-custom-options to magit-key-mode.el
It should probably also be renamed to something with a "magit-key-mode-"
prefix.  But another option (which I prefer) would be to use loose the
useless distinction between it and `magit-key-mode-current-options' so
I don't do it (at least for now).

Because `magit-key-mode' is not required until the end of "magit.el" the
variable now also has to be declared in the latter.  I intend to change
that so this is only a tempory hack.