drop support for Git versions older than 1.7.2.5
This was proposed in #709 and also mentioned on the mailing list.
There have been no objects.

Version 1.7.2.5 is now the oldest supported version.  That is the
version available in Debian oldstable (6.0, Squeeze).

Remove the `magit-configure-have-{*}' functions and `magit-have-{*}'
variables that were used to disable certain functionality for older
Git versions.

In areas affected by this change also sort Git options so that they
are grouped, as much as possible, according to the sections in the
`git-log' man page.