use file-local variables instead of dir-locals.el
To not use the file `.dir-locals.el' to set buffer local variables for
Emacs lisp libraries.  Instead set `indent-tabs-mode' to nil in each
library individually.  That leads to a bit of duplication, but has the
advantage, that users can now use `.dir-locales.el' for their own
settings, without that file being tracked by git.  To make that even
more convenient tell git to ignore it.