magit-key-mode.el: lose many doc-strings
In `magit-key-mode.el' lose the doc-strings of many variables and
functions.  The reasons vary but generally it is done because the
existing doc-strings are either insufficient or document things that
are about to change or to be removed.  Better doc-strings will be
added once the current refactoring is over, for things that actually
need to be documented.  Removing the old doc-strings now makes the
refactoring easier, and writing new ones will also be easier without
"inspiration" from not so great ones.

The only doc-strings that are kept are those of faces and those of
`magit-define-popup' forms.

Also shorten Commentary (to be extended later) and lose some
unnecessary comments.