popups: use terms argument, option, switch, and action consistently
Arguments come in two forms, switches and options.
Switches are arguments that can be toggled on or off.
Options are arguments that take a value.

Actions are commands whose invocation cause an actual change, as opposed
to setting the value of an argument, which is also implemented using
Emacs commands but only arranges for the subsequent invocation of an
action to have a different effect.

In the future there will also be "positional arguments" but currently
the popup code does not handle these.  Currently some action commands
always prompt the user for some minibuffer input, while others only do
so when a prefix argument (only loosely related to "argument" as defined
above) are used.

Do *not* define an alias for the renamed `magit-key-mode-args-face'; if
users and theme authors did take inspiration from the default look, then
we *want* to break that, as the text propertized with this face is not
editable the same way a `widget-field' is.  The actual face definition
still has to be adapted accordingly.