popups: define popups next to their commands
Define popups (using the new `magit-define-popup') right before the
definitions of the commands that implement the "actions" that can be
invoked from the popup.  This separates the functionality implemented
in `magit-key-mode.el' from its uses, and as a result this library and
`magit.el' no longer depend on one another.  `magit-key-mode.el' now
does not depend on anything else in Magit, and can theoretically be
used outside of Magit.