magit-this-popup: new variable
Instead of passing the symbol that identifies the current popup around
as an argument, store it in the buffer-local variable `magit-this-popup'.

This allows turning `magit-popup-help' into a proper command, instead of
having to wrap it with an interactive lambda expression in each popup's
keymap.  In the next commit we will really benefit from this change, by
doing the same for other such lambda commands, effectively making popup
specific keymaps unnecessary.