magit-define-popup-{*}: use for customization
Previously the `magit-define-popup-switch', `magit-define-popup-option',
and `magit-define-popup-action' were used by `magit-define-popup-keymap'
to create a keymap from a popup definition.  That is now done directly
in the latter.

These functions are now repurposed to let users customize an existing
popup definition.  In other words they now serve a similar purpose as
`define-key' does for regular keymaps.

Because each popup is still stored in two places, once as keymap and
once as an alist, these functions are still quite ugly.  That will be
fixes soon, when we stop using a keymap per popup.