popups: refactor invocation of popup events
The macro `magit-define-popup' is used to define a popup, i.e. a command
to insert the popup; the arguments that can be set in the popup; and the
commands that can be used to exit the popup, using the set arguments.

Prior to this commit that meant defining a Emacs command and *three*
variables; one variable uses the same name as the command and is used to
store the popup plist.  Another is a customizable variable used to store
default values for the arguments.  And finally there was a variable that
used to store the popup specific keymap.

This commit removes the need for that keymap variable.  Now there is
only one keymap which uses remapping and default (`t') bindings to bind
the commands `magit-invoke-popup-action', `magit-invoke-popup-switch,
`magit-invoke-popup-option'.  Prior to this commit these were not Emacs
commands but non-interactive functions, and for each "binding" in each
popup an anonymous lambda command had to be defined.  Now they are
commands and they use information stored in the popup variable to do key
lookup.

* Add function `magit-invoke-popup' which is now called by each popup
  entry command generated by `magit-define-popup'.  Previously
  `magit-popup-mode-setup' did fill that role (and currently the former
  does nothing more than calling the latter).

* Add buffer-local variable `magit-this-popup-events' replacing
  `magit-popup-args'.  We now need to store actions in here too because
  we want to switch to an internal format, different from the alist
  stored in the "definition variable".

* Add function `magit-popup-get' which works like `plist-get' but
  knows whether the requested property is stored in the "definition
  variable" or in `magit-this-popup-events'.  It's also less verbose
  to use because the PLIST argument is omitted.

* Add function `magit-popup-get-args' which returns the list of
  arguments the way they are exposed as `magit-current-popup-args'.

* Add function `magit-popup-setup-events' which converts information
  from the "definition variable" to the internal format.  Currently
  it looks simple enough, unfortunately that will change as the
  internal format get's more complex.

* Remove function `magit-define-popup-keymap', because, there is but
  one keymap.

* Promote `magit-invoke-popup-{*}' to commands.  They should not be
  used outside a popup.

* Let-bind `magit-current-popup' when invoking an action from a popup.
  This is similar to `magit-current-popup-args' but holds the symbol
  identifying the popup, as opposed to the selected arguments.  This
  might be useful for actions that can be invoked from multiple
  popups.