popups: handle "q" event explicitly
An event explicitly bound in `magit-mode-map' cannot be bound to an
action without explicitly using a popup local keymap that overrides
that binding.  Because all letters should be available for actions,
stop binding q to `magit-popup-quit' and instead explicitly call that
command in `magit-invoke-popup-action' if no action uses it.  We have
to do so because Magit users are used to using q to quit the popup.