popups: add support for default actions
For each popup, define a default action.

Add option `magit-popup-use-prefix-argument' which controls when the
default action is used.

Because the old implementation used the prefix argument in an
incompatible fashion and changing that without users being aware of
it could lead to tears the default is this:

Don't show the popup and also don't use the default action, instead
make the user aware of the alternatives and make her choose one:

* Without a prefix argument show the popup,
  with a prefix argument directly invoke the default action.

* With a prefix argument show the popup,
  without a prefix argument directly invoke the default action.

* Ignore the prefix argument and always show the popup.