popups: use list of strings for all popup options
Previously the value of "popup default arguments" custom options were
alists, now they are lists of strings, i.e arguments that can be
passed to git without any pre-processing.

Using an alist was a bad idea because (1) we cannot do the same when
the value is used directly (what the "setvar" mode does), (2) and as
a result of that a lot of code had to exist in two variants.

It would have been better to remove this completely from history, but
reverting this feature when I added it would result in to many commits
with difficult merge conflicts.  Sorry in this case history reflects
what I have actually done, not what I came to know I should have done :-)