add library magit-popup
... or rather rename library `magit-key-mode' to `magit-popup'.

During the last ~90 commits the popup functionality was completely
refactored, and renaming the file is the final step in replacing the
old library with the new implementation.

Declare that I am the author of `magit-popup.el' but add a prominent
note to the library header about it being inspired by Phil Jackson's
`magit-key-mode'.

Without the latter there would also be no `magit-popup.el'.  On the
other hand there is virtually nothing left from `magit-key-mode.el',
in the new implementation.

Also adjust the library summary and add a library commentary.