temporarily remove "update diff arguments" functionality
Previously the same popup was used to show a "new diff with the selected
arguments" and to "update the arguments used by the current diff", which
was rather confusing.  A bigger issue however is that in the former case
the arguments are not actually respected.  To make it easier to fix that
first remove the ability to do the latter.  Once the former has been
fixed, the functionality removed here will also be restored, but then
using a dedicated popup and new setter commands.