popup: fix inserting gaps
The problem was that we inserted a gap by *completely* filling the
column with spaces.  Padding happens when inserting the *next* column
and because buttons *never* are long enough to fill the column, the
code responsible for doing that came to the conclusion we are about
to insert column n+1 when we were about to insert column n.  It then
noticed that the space reserved for column n was not completely used
up (actually not at all) and inserted spaces for that column.

By inserting a single space we fix that.  Inserting nothing at all for
a gap would not work, because then we would have the same problem in
reverse.  The code responsible for padding would then, when inserting
the next item think we are still at column n-1 when we actually are at
column n.