popup: fix edge case related to push-button
Don't do anything when the user actually pushes the button labeled
"Push button".  Otherwise we would enter a loop.  Not doing anything
is okay; that button is only intended as "documentation", not for
touching.  Would sensible thing could a user possibly expect to
happen when pushing a button labeled "Push button"?