use async-bytecomp to guarantee untainted builds
When `package.el' is used to update a package whose libraries have
already been loaded in the current session, then parts of the loaded
version can leak into the new version during byte-compilation.

This can be avoided by compiling the libraries in a separate Emacs
instance.  The library `async-bytecomp', which is part of the package
`async', does just that by advising the function `package--compile'.

So add `async' as an Elpa dependency, soft-require `async-bytecomp', and
if that succeeds, then turn on `async-bytecomp-package-mode' to activate
the advice.  Because it is undefined which package is updated first, and
which library is loaded first, we have to this in both possible
candidates `with-editor' and `magit-popup'.

Users may update `magit' without updating the dependencies, so we have
to further force the updates of all the Elpa packages that are build
from the Melpa repository.  To do so we depend on "20150808" snapshot
versions instead of "2.2.0".

When installing Magit manually, then `async-bytecomp' is not required,
because Magit is then compiled using the Makefile which already uses a
separate, untainted Emacs instance.