popup: support new :variable event type
Add support for a new popup event type `:variable' to change Git
variables from a popup.  `magit-popup.el' isn't really up for the
challenge, so this is a bit of a hack.  It's intentional that the
documentation is sparse, as I don't want to encourage use outside
of Magit.  Magit is required to be able to use this new type, but
the byte-compiler won't complain if it is not available.