Add support for JќL key bindings
JќL is the name I have given to my (for now) personal key bindings,
which replace "bpnf" with "jikl" for navigation (similar to how vi
uses "hjkl", but non-modal).  Maintaining such changed key bindings
across all the packages I use has proven to be challenging to say the
least. Recently I have decided to try a new approach - just patch the
packages that I use.  That might seem like bad idea, and maybe it is,
but all the others didn't really scale either.  (And considering that
I have written a package manager, Epkg, that is supposed to help with
exactly this sort of thing, I ought to try that approach.)

However for Magit patching is not feasible, because I work on that all
the time, and sometimes when I work on a feature branch I do have to
restart Emacs.  I need my key bindings even when I am on a feature
branch and not by own personal key bindings branch.

Sometimes I add some optional functionality to Magit for the benefit
of a very small number of users, sometimes consisting of a single
user, who manages to convince me how very important a change is, even
though nobody else seems to have the same need.  Well, this time that
user is me.