manual: add dummy subsection, wrapping initial section content
This makes the CONTENTS visibility state much more useful.  Now that
state shows nothing but headings, including the newly added pseudo
headings "_ :ignored:".  Previously the initial text belonging to a
section that also has subsections, was also visible in this state.

These "_ :ignored:" headers are not exported to `texinfo' -- that's
the "pseudo" part.  This filtering is implemented in the `ox-extra'
package, which is distributed with Org.  Visiting the Org source of
one of our manuals automatically enables this feature.

Unfortunately the filtering would result in duplicated menus being
inserted when using the `texinfo' exporter, so Kyle wrote an advice
for `org-export-ignore-headlines', which this commit places in
"magit-utils.el".  Hopefully we will eventually be able to remove
this hack.  See #2914 to keep track of this.