magit-dwim-selection: new option
`magit-invoke-popup-action' is adjusted to set `this-command' to the
invoked command before calling instead of after doing so.  Without
this change `this-command's value would be wrong when consulting it
in `magit-completing-read' when the command was invoked form a popup.

Also deprecate the option `magit-no-confirm-default', which will be
removed in v2.13.0.

Re #2916.

squash! magit-dwim-selection: new option