No longer depend on async package
We used to do that so that magit-popup would get properly recompiled
even when there are changes to macros and classes.  We don't need to
do that anymore because we are in maintenance mode and don't make such
changes anymore.  We remove this because it has bit-rotten; the format
of `async-bytecomp-allowed-packages' has changed, see #15.